Moved other mods text stuff to a separate file
