--- sadx-russian-mod/BetaRestores.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void OverwriteBetaRestoresText(const HelperFunctions& helperFunctions);
