+++ sadx-russian-mod/OtherMods.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void OverwriteBetaRestoresText(const HelperFunctions& helperFunctions);
+void ReplaceMultiplayerText(const HelperFunctions& helperFunctions);
+void ReplaceSuperSonicHint();
