+++ sadx-russian-mod/sadx-russian-mod.vcxproj
@@ -79,7 +79,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="BetaRestores.cpp" />
+    <ClCompile Include="OtherMods.cpp" />
     <ClCompile Include="cutscene-decomp\Amy\ev0060_a_intro1.cpp" />
     <ClCompile Include="cutscene-decomp\Amy\ev0061_a_intro2.cpp" />
     <ClCompile Include="cutscene-decomp\Amy\ev0062_a_meetsonic.cpp" />
@@ -122,7 +122,6 @@
     <ClCompile Include="VariousText.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="BetaRestores.h" />
     <ClInclude Include="cutscene-decomp\include\SADXEventFunctions.h" />
     <ClInclude Include="cutscene-decomp\include\SADXEventList.h" />
     <ClInclude Include="cutscene-decomp\include\SADXEventVariables.h" />
@@ -130,6 +129,7 @@
     <ClInclude Include="Macros.h" />
     <ClInclude Include="LoadTextures.h" />
     <ClInclude Include="ModConfig.h" />
+    <ClInclude Include="OtherMods.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="CustomSubTimings.h" />
     <ClInclude Include="VariousText.h" />
